web_app/requirements.txt
absl-py==0.11.0
astor==0.8.1
astunparse==1.6.3
attrs==20.3.0
beautifulsoup4==4.9.3
bert-serving-client==1.10.0
bert-serving-server==1.10.0
bleach==1.5.0
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Click==7.0
dataclasses==0.8
dill==0.3.3
Flask==1.0.2
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
GPUtil==1.4.0
grpcio==1.33.2
h5py==2.10.0
html5lib==0.9999999
idna==2.10
ImageHash==4.1.0
importlib-metadata==2.0.0
importlib-resources==3.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.2
Pillow==6.2.2
promise==2.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytesseract==0.2.6
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyzmq==20.0.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
soupsieve==2.0.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-datasets==4.1.0
tensorflow-estimator==2.3.0
tensorflow-metadata==0.25.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
tqdm==4.52.0
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==0.16.0
wrapt==1.12.1
zipp==3.4.0
torch==1.7.0
transformers==2.5.1